+++ tests/testthat/test-errors.R
@@ -37,5 +37,25 @@ test_that("errors can be defined as integers", {
 test_that("type_sum is available for errors objects", {
   skip_if_not_installed("tibble")
   library(tibble)
-  expect_equal(type_sum(set_errors(1, 0.1)), "errors")
+  expect_equal(type_sum(set_errors(1, 0.1)), "[(err)]")
+})
+
+test_that("pillar_shaft is available for errors objects", {
+  skip_if_not_installed("pillar")
+  library(pillar)
+  expect_equal(as.character(pillar_shaft(set_errors(1, 0.1))),
+               paste0("1.0", style_subtle("(1)")))
+})
+
+test_that("errors are dropped", {
+  x1 <- set_errors(1:3, 0.1)
+  x2 <- set_errors(x1, NULL)
+  x3 <- drop_errors(x1)
+  errors(x1) <- NULL
+  expect_true(!inherits(x1, "errors"))
+  expect_true(!inherits(x2, "errors"))
+  expect_true(!inherits(x3, "errors"))
+  expect_null(attr(x1, "errors"))
+  expect_null(attr(x2, "errors"))
+  expect_null(attr(x3, "errors"))
 })