+++ NEWS.md
@@ -1,3 +1,19 @@
+# errors 0.2.1
+
+## New features
+
+* New `drop_errors` method drops `errors` class and attributes. Equivalent to setting a `NULL` error (#23).
+* Implement `pillar_shaft` for tidy `tibble` printing and improve `type_sum` (700f1aa).
+
+## Minor changes and fixes:
+
+* Show `Ops` warnings once (#22).
+* Refactor `print` method (2c252d9).
+* Drop support for matrix multiplication: no warning is issued anymore (090e953).
+* Fix `cbind`/`rbind` column/row name parsing for named arguments (b8ea6b5).
+* Improve plus-minus separator (8b3d231).
+* Fix formatting to take into account the `scipen` option (9299551).
+
 # errors 0.2.0
 
 ## Minor changes and fixes: